+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add `retried` to `Transaction`
+
 ## 4.14.0
 * Add `PaymentMethodCustomerDataUpdated` webhook